+++ src/Utility/BackupManager.php
@@ -25,6 +25,7 @@ use InvalidArgumentException;
 use Symfony\Component\Finder\Finder;
 use Symfony\Component\Finder\SplFileInfo;
 use Tools\Exceptionist;
+use Tools\Filesystem;
 
 /**
  * Utility to manage database backups
@@ -37,48 +38,47 @@ class BackupManager
      * Deletes a backup file
      * @param string $filename Filename of the backup that you want to delete.
      *  The path can be relative to the backup directory
-     * @return bool
+     * @return string
      * @see https://github.com/mirko-pagliai/cakephp-database-backup/wiki/How-to-use-the-BackupManager-utility#delete
      * @throws \Tools\Exception\FileNotExistsException
      * @throws \Tools\Exception\NotWritableException
      */
-    public function delete(string $filename): bool
+    public static function delete(string $filename): string
     {
-        $filename = $this->getAbsolutePath($filename);
+        $filename = self::getAbsolutePath($filename);
+        Filesystem::instance()->remove(Exceptionist::isWritable($filename));
 
-        return unlink(Exceptionist::isWritable($filename));
+        return $filename;
     }
 
     /**
      * Deletes all backup files
-     * @return array List of deleted backup files
+     * @return array<string> List of deleted backup files
      * @see https://github.com/mirko-pagliai/cakephp-database-backup/wiki/How-to-use-the-BackupManager-utility#deleteAll
      * @since 1.0.1
      * @uses delete()
      * @uses index()
      */
-    public function deleteAll(): array
+    public static function deleteAll(): array
     {
-        return array_filter(array_map(function (string $filename) {
-            return !$this->delete($filename) ?: $filename;
-        }, $this->index()->extract('filename')->toList()));
+        return array_map('self::delete', self::index()->extract('filename')->toList());
     }
 
     /**
      * Returns a list of database backups
-     * @return \Cake\Collection\CollectionInterface Collection of backups. Each backup
-     *  is an entity
+     * @return \Cake\Collection\CollectionInterface<\Cake\ORM\Entity> Backups
      * @see https://github.com/mirko-pagliai/cakephp-database-backup/wiki/How-to-use-the-BackupManager-utility#index
      */
-    public function index(): CollectionInterface
+    public static function index(): CollectionInterface
     {
         $finder = (new Finder())->files()->name('/\.sql(\.(gz|bz2))?$/')->in(Configure::read('DatabaseBackup.target'));
 
         return collection($finder)->map(function (SplFileInfo $file) {
-            return new Entity([
-                'filename' => $file->getFilename(),
-                'extension' => $this->getExtension($file->getFilename()),
-                'compression' => $this->getCompression($file->getFilename()),
+            $filename = $file->getFilename();
+
+            return new Entity(compact('filename') + [
+                'extension' => self::getExtension($filename),
+                'compression' => self::getCompression($filename),
                 'size' => $file->getSize(),
                 'datetime' => FrozenTime::createFromTimestamp($file->getMTime()),
             ]);
@@ -91,17 +91,17 @@ class BackupManager
      * You must indicate the number of backups you want to keep. So, it will
      *  delete all backups that are older.
      * @param int $rotate Number of backups that you want to keep
-     * @return array Array of deleted files
+     * @return array<\Cake\ORM\Entity> Array of deleted files
      * @see https://github.com/mirko-pagliai/cakephp-database-backup/wiki/How-to-use-the-BackupManager-utility#rotate
      * @throws \InvalidArgumentException
      * @uses delete()
      * @uses index()
      */
-    public function rotate(int $rotate): array
+    public static function rotate(int $rotate): array
     {
         Exceptionist::isPositive($rotate, __d('database_backup', 'Invalid rotate value'), InvalidArgumentException::class);
-        $backupsToBeDeleted = $this->index()->skip((int)$rotate);
-        array_map([$this, 'delete'], $backupsToBeDeleted->extract('filename')->toList());
+        $backupsToBeDeleted = self::index()->skip((int)$rotate);
+        array_map('self::delete', $backupsToBeDeleted->extract('filename')->toList());
 
         return $backupsToBeDeleted->toList();
     }
@@ -115,11 +115,10 @@ class BackupManager
      * @since 1.1.0
      * @throws \Tools\Exception\NotReadableException
      */
-    protected function getEmailInstance(string $backup, string $recipient): Email
+    protected static function getEmailInstance(string $backup, string $recipient): Email
     {
-        $file = $this->getAbsolutePath($backup);
-        Exceptionist::isReadable($file);
-        $basename = basename($file);
+        $file = self::getAbsolutePath($backup);
+        $basename = basename(Exceptionist::isReadable($file));
         $server = env('SERVER_NAME', 'localhost');
 
         return (new Email())
@@ -140,6 +139,6 @@ class BackupManager
      */
     public function send(string $filename, string $recipient): array
     {
-        return $this->getEmailInstance($filename, $recipient)->send();
+        return self::getEmailInstance($filename, $recipient)->send();
     }
 }